% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{.rustgd_frames_bin}
\alias{.rustgd_frames_bin}
\title{Internal: absolute path to the staged rustgd-frames binary.}
\usage{
.rustgd_frames_bin()
}
\description{
Internal: absolute path to the staged rustgd-frames binary.
}
