% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{.rustgd_frames_register_cleanup}
\alias{.rustgd_frames_register_cleanup}
\title{Internal: remove the frames directory when R exits.}
\usage{
.rustgd_frames_register_cleanup()
}
\description{
Internal: remove the frames directory when R exits.
}
