% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rustgd_set_current_page}
\alias{rustgd_set_current_page}
\title{Set the current page number directly. R-side uses this during
multi-plot history replay to direct each replayed plot's flushes
to a specific plot-NNNN.svg file. Only meaningful in conjunction
with replay mode.}
\usage{
rustgd_set_current_page(page)
}
\description{
Set the current page number directly. R-side uses this during
multi-plot history replay to direct each replayed plot's flushes
to a specific plot-NNNN.svg file. Only meaningful in conjunction
with replay mode.
}
