% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webview.R
\name{rustgd_webview_bin}
\alias{rustgd_webview_bin}
\title{Internal: absolute path to the staged rustgd-webview binary.}
\usage{
rustgd_webview_bin()
}
\description{
Internal: absolute path to the staged rustgd-webview binary.
}
